feat: implement Habitacion microservice with controller, service, repository, mapper, and security configuration updates

+++ commons-hotel/src/main/java/com/hotelreservation/commons/dto/habitaciones/HabitacionRequest.java
@@ -19,7 +19,6 @@ public record HabitacionRequest(
 
         @NotNull(message = "La capacidad es obligatoria")
         @Min(value = 1, message = "La capacidad mínima es 1")
-        Integer capacidad
-
+        Short capacidad
 ) {
 }
