+++ commons-hotel/src/main/java/com/hotelreservation/commons/dto/habitaciones/HabitacionResponse.java
@@ -7,6 +7,6 @@ public record HabitacionResponse(
         Integer numero,
         String tipo,
         BigDecimal precio,
-        Integer capacidad
+        Short capacidad
 ) {
 }
